+++ src/nodejs/lambda-handlers/file-upload.js
@@ -148,8 +148,8 @@ const operations = {
 };
 
 async function handler(event) {
-  return { error: 'Not Implemented' };
-  /* eslint-disable no-unreachable */
+  /* return { error: 'Not Implemented' };
+  eslint-disable no-unreachable */
   console.info(`[EVENT]\n${JSON.stringify(event)}`);
   const user = event.context.user_id;
   const operation = operations[event.operation];